cignaFastAPI/requirements.txt
fastapi
uvicorn
boto3==1.20.23
requests==2.26.0
tqdm
python-dotenv==0.19.2